pyproject.toml
[project]
name = "rag-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-core>=1.3.0",
    "langchain-ollama>=1.1.0",
    "ollama>=0.6.1",
    "pandas>=3.0.2",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.26",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.6.0",
    "unstructured>=0.22.22",
    "uvicorn>=0.45.0",
]
